code review - move code in wizard to hook useWizard
