+++ src/Routes/SidebarRight/AddPipeline/Wizard.js
@@ -1,23 +1,17 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-
 import { CheckOutlined, RightOutlined, LeftOutlined } from '@ant-design/icons';
-
 import { Steps, Form as AntdForm } from 'antd';
 import { JsonView } from 'components/common';
 import styled from 'styled-components';
 import { COLOR_LAYOUT } from 'styles';
-import { pickBy, identity } from 'lodash';
 import {
   BottomPanel,
   PanelButton,
   RightAlignedButton,
 } from 'components/Drawer';
-import { Initial, Nodes, Options } from './Steps';
+import { useWizard } from 'hooks';
 import { context } from './useWizardContext';
-import useSubscribe from '../useSubscribe';
-
-const pruneObject = obj => pickBy(obj, identity);
 
 const Form = styled(AntdForm)`
   width: 90ch;
@@ -30,36 +24,6 @@ export const Body = styled.div`
   overflow-y: scroll;
   max-height: 81vh;
 `;
-
-const stepNames = ['Initial', 'Nodes', 'Options']; // 'Nodes', 'Options'
-const stepComponents = [Initial, Nodes, Options];
-
-const steps = stepNames.map(name => (
-  <Steps.Step key={`steps-${name}`} title={name} />
-));
-
-/** @param {[any]} collection */
-const normalize = collection =>
-  collection.reduce((acc, item, ii) => ({ ...acc, [ii]: item }), {});
-
-// converts arrays to objects on selected fields to match ant requirement
-// eslint-disable-next-line no-unused-vars
-const parseInitialState = initialState => {
-  const nodes =
-    initialState?.nodes?.map(item =>
-      !item.input
-        ? item
-        : {
-            ...item,
-            input: normalize(item.input),
-          }
-    ) ?? [];
-  const state = {
-    ...initialState,
-    nodes: normalize(nodes),
-  };
-  return state;
-};
 
 /** @param {object} props */
 /** @param {import('antd/lib/form').FormProps} props.form */
@@ -74,87 +38,28 @@ const Wizard = ({
   wizardClear,
   isEdit,
 }) => {
-  const [valuesState, setValuesState] = useState({});
-  const { setFieldsValue, getFieldsValue, getFieldValue } = form;
-  const { subscribe } = useSubscribe();
-
-  useEffect(() => {
-    setFieldsValue(pruneObject(initialState));
-  }, [setFieldsValue, initialState]);
-
-  const getFormattedFormValues = useCallback(() => {
-    const formValues = getFieldsValue();
-    delete formValues.listKeyValue;
-
-    const nodes = Object.values(formValues?.nodes || {})
-      .filter(item => item?.kind)
-      .map(item => {
-        if (!item?.input) return item;
-        return {
-          ...item,
-          input: Object.values(item?.input),
-        };
-      });
-
-    return pruneObject({ ...formValues, nodes });
-  }, [getFieldsValue]);
-
-  const persistForm = useCallback(() => {
-    setEditorState(getFormattedFormValues());
-  }, [setEditorState, getFormattedFormValues]);
-
-  useEffect(() => subscribe(persistForm), [subscribe, persistForm]);
-
-  const isLastStep = stepIdx === steps.length - 1;
-
-  const onPrevious = useCallback(() => setStepIdx(state => state - 1), [
+  const {
+    steps,
+    setForm,
+    handleSubmit,
+    isStreamingPipeline,
+    valuesState,
+    stepComponents,
+    stepNames,
+    getFormattedFormValues,
+    handleToggle,
+    onPrevious,
+    isLastStep,
+    onNext,
+  } = useWizard(
+    form,
+    initialState,
+    stepIdx,
     setStepIdx,
-  ]);
-
-  const onNext = useCallback(() => setStepIdx(state => state + 1), [
-    setStepIdx,
-  ]);
-
-  const handleToggle = useCallback(() => {
-    persistForm();
-    toggle();
-  }, [persistForm, toggle]);
-
-  const handleSubmit = useCallback(
-    e => {
-      e?.preventDefault();
-      onSubmit(getFormattedFormValues());
-    },
-    [getFormattedFormValues, onSubmit]
+    toggle,
+    onSubmit,
+    setEditorState
   );
-  // check for undefined to avoid removing streaming only fields while initial load
-  const isStreamingPipeline = ['stream', undefined].includes(
-    getFieldValue('kind')
-  );
-
-  const resetKind = useCallback(
-    typeKind => {
-      const { nodes } = getFieldsValue();
-      nodes &&
-        nodes.forEach((node, index) => {
-          if (node?.kind === typeKind) {
-            nodes[index].kind = 'algorithm';
-            setFieldsValue({ nodes });
-          }
-        });
-    },
-    [getFieldsValue, setFieldsValue]
-  );
-
-  useEffect(() => {
-    // remove gateway or output option from nodes and reset them to algorithm option
-    if (isStreamingPipeline) resetKind('output');
-    else resetKind('gateway');
-  }, [isStreamingPipeline, resetKind]);
-
-  const setForm = useCallback(() => {
-    setValuesState(getFormattedFormValues());
-  }, [getFormattedFormValues, setValuesState]);
 
   return (
     <>
