+++ src/hooks/index.js
@@ -25,6 +25,7 @@ export { default as useTraceData } from './useTraceData';
 export { default as useVersions } from './useVersions';
 export { default as useSiteDarkMode } from './useSiteDarkMode.react';
 export { default as useWizardAddPipeline } from './useWizardAddPipeline';
+export { default as useWizard } from './useWizard';
 
 export const dataSources = {
   useSnapshots: useDataSourceSnapshots,
